General: removed some dependencies to GLFW where possible, added std:: prefix to not expect `using namespace std` in magic places; Root: fixed some subtle bugs surrounding popup handling, fix-up the double tap property; ScrollPane: new option to overlay another panel above the main view; SegmentedSelect: fixed nullpointer dereference in setSelectedIndex(), Tweener: new findTween(id) and removeParameter(float*) methods, user-defined identifier field on tweens

+++ src/Label.cpp
@@ -88,7 +88,7 @@ void mui::Label::commit(){
         if( boundingBox.width > width && text.length() > 3 ){
 			// don't mess with the original boundingBox object. it's publicly readable!
 			ofRectangle box = boundingBox;
-            ssize_t len = (ssize_t)text.length() - 3;
+            int64_t len = (int64_t)text.length() - 3;
             while( box.width > width && len >= 0 ){
                 displayText = text.substr(0, len ) + "...";
 				box = Helpers::getFontStash().getTextBounds(displayText, fontStyle, 0, 0 );
